+++ packages/argdown-parser/lib/test/lexer.spec.js
@@ -346,7 +346,7 @@ describe("Lexer", function () {
         var source = _fs2.default.readFileSync("./test/lexer-windows-line-endings.argdown", "utf8");
         var result = lexer.tokenize(source);
         startTest(result.tokens);
-        console.log(lexer.tokensToString(result.tokens));
+        //console.log(lexer.tokensToString(result.tokens));
         //expect(result.tokens.length).to.equal(5);
         expectToken(lexer.Freestyle);
         expectToken(lexer.Indent);